python_ai/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-Login==0.6.2
Flask-CORS==4.0.0
Pillow==10.0.0
pytesseract==0.3.10
PyMuPDF==1.23.8
sentence-transformers==2.7.0
numpy==1.26.4
python-docx==0.8.11
ebooklib==0.19.0
nltk==3.8.1
rapidfuzz==2.15.1
pdfplumber==0.10.0
scikit-learn==1.4.0
torch==2.0.0





